requirements.txt
Appium-Python-Client
allure-python-commons==2.8.7
allure-robotframework==2.8.7
attrs==18.2.0
dpath==2.0.1
Pillow==5.4.1
pluggy==0.8.1
requests==2.22.0
robotframework==3.1.2
selenium==3.141.0
visual-logging==1.0
webdriver-manager==2.3.0
